moved from int to int64_t for id params

+++ src/back/project/src/model/cc_di.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct CcDi {
-	int ccId;
-	int diId;
+	int64_t ccId;
+	int64_t diId;
 };
 
 } // namespace svetit::project::model
