+++ src/back/project/src/model/cc_param.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct CcParam {
-	int ccId;
-	int paramId;
+	int64_t ccId;
+	int64_t paramId;
 };
 
 } // namespace svetit::project::model
