+++ src/back/project/src/model/cc_type_di_type.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct CcTypeDiType {
-	int ccTypeId;
-	int diTypeId;
+	int64_t ccTypeId;
+	int64_t diTypeId;
 };
 
 } // namespace svetit::project::model
