+++ src/back/project/src/model/cc_type_param.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct CcTypeParam {
-	int ccTypeId;
-	int paramId;
+	int64_t ccTypeId;
+	int64_t paramId;
 };
 
 } // namespace svetit::project::model
