+++ src/back/project/src/model/cc_type_serialize.cpp
@@ -29,7 +29,7 @@ CcType Parse(
 	const auto projectId = projectIdStr.empty() ? boost::uuids::uuid{} : utils::BoostUuidFromString(projectIdStr);
 
 	return {
-		.id = json["id"].As<int>(),
+		.id = json["id"].As<int64_t>(),
 		.projectId = projectId,
 		.key = json["key"].As<std::string>(),
 		.name = json["name"].As<std::string>(),
