+++ src/back/project/src/model/device_plugin_param_serialize.cpp
@@ -21,8 +21,8 @@ DevicePluginParam Parse(
 	formats::parse::To<DevicePluginParam>)
 {
 	return {
-		.deviceId = json["deviceId"].As<int>(),
-		.paramId = json["paramId"].As<int>()
+		.deviceId = json["deviceId"].As<int64_t>(),
+		.paramId = json["paramId"].As<int64_t>()
 	};
 }
