+++ src/back/project/src/model/di_plugin_param.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstdint>
+
 namespace svetit::project::model {
 
 struct DiPluginParam {
-	int diTypeId;
-	int paramId;
+	int64_t diTypeId;
+	int64_t paramId;
 };
 
 } // namespace svetit::project::model
