+++ src/back/project/src/model/section_serialize.cpp
@@ -27,7 +27,7 @@ Section Parse(
 	const auto projectId = projectIdStr.empty() ? boost::uuids::uuid{} : utils::BoostUuidFromString(projectIdStr);
 
 	return {
-		.id = json["id"].As<int>(),
+		.id = json["id"].As<int64_t>(),
 		.projectId = projectId,
 		.name = json["name"].As<std::string>()
 	};
