+++ src/back/project/src/model/value_view.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 namespace svetit::project::model {
 
 struct ValueView {
-	int id;
-	int diTypeId;
+	int64_t id;
+	int64_t diTypeId;
 	std::string value;
 	std::string view;
 };
