+++ src/back/project/src/repo/table_cc_mode_type.hpp
@@ -12,10 +12,10 @@ namespace svetit::project::table {
 class CcModeType final {
 public:
 	explicit CcModeType(storages::postgres::ClusterPtr pg);
-	model::CcModeType Get(int id);
+	model::CcModeType Get(int64_t id);
 	int64_t Create(const model::CcModeType& ccModeType);
 	void Update(const model::CcModeType& ccModeType);
-	void Delete(int id);
+	void Delete(int64_t id);
 	PagingResult<model::CcModeType> GetList(int start, int limit);
 private:
 	storages::postgres::ClusterPtr _pg;
