+++ src/back/project/src/repo/table_cc_status_type.cpp
@@ -21,7 +21,7 @@ const pg::Query kGet{
 	pg::Query::Name{"select_cc_status_type"},
 };
 
-model::CcStatusType CcStatusType::Get(int id) {
+model::CcStatusType CcStatusType::Get(int64_t id) {
 	auto res = _pg->Execute(ClusterHostType::kMaster, kGet, id);
 	if (res.IsEmpty())
 		throw errors::NotFound404{};
@@ -59,7 +59,7 @@ const pg::Query kDelete {
 	pg::Query::Name{"delete_cc_status_type"},
 };
 
-void CcStatusType::Delete(int id) {
+void CcStatusType::Delete(int64_t id) {
 	auto res = _pg->Execute(ClusterHostType::kMaster, kDelete, id);
 	if (!res.RowsAffected())
 		throw errors::NotFound404();
