+++ src/back/project/src/repo/table_cc_type.hpp
@@ -12,10 +12,10 @@ namespace svetit::project::table {
 class CcType final {
 public:
 	explicit CcType(storages::postgres::ClusterPtr pg);
-	model::CcType Get(int id);
+	model::CcType Get(int64_t id);
 	int64_t Create(const model::CcType& ccType);
 	void Update(const model::CcType& ccType);
-	void Delete(int id);
+	void Delete(int64_t id);
 	PagingResult<model::CcType> GetList(int start, int limit);
 private:
 	storages::postgres::ClusterPtr _pg;
