+++ src/back/project/src/repo/table_cc_type_di_type.hpp
@@ -12,9 +12,9 @@ namespace svetit::project::table {
 class CcTypeDiType final {
 public:
 	explicit CcTypeDiType(storages::postgres::ClusterPtr pg);
-	model::CcTypeDiType Get(int ccTypeId, int diTypeId);
-	void Insert(int ccTypeId, int diTypeId);
-	void Delete(int ccTypeId, int diTypeId);
+	model::CcTypeDiType Get(int64_t ccTypeId, int64_t diTypeId);
+	void Insert(int64_t ccTypeId, int64_t diTypeId);
+	void Delete(int64_t ccTypeId, int64_t diTypeId);
 	PagingResult<model::CcTypeDiType> GetList(int start, int limit);
 private:
 	storages::postgres::ClusterPtr _pg;
